+++ src/test/java/net/kaczmarzyk/spring/data/jpa/ApplicationWithConfiguredConversionService.java
@@ -38,7 +38,8 @@ import static org.springframework.jmx.support.RegistrationPolicy.IGNORE_EXISTING
  */
 @Configuration
 @ComponentScan(basePackages = "net.kaczmarzyk", excludeFilters = {
-		@ComponentScan.Filter(type = ASSIGNABLE_TYPE, value = Application.class)
+		@ComponentScan.Filter(type = ASSIGNABLE_TYPE, value = Application.class),
+		@ComponentScan.Filter(type = ASSIGNABLE_TYPE, value = ApplicationWithSARConfiguredWithApplicationContext.class)
 })
 @EnableJpaRepositories
 @EnableAutoConfiguration